Reorganize base code
