+++ src/pykeen/datasets/hyper/__init__.py
@@ -1,0 +1,6 @@
+from .base import HyperRelationalDataset, HyperRelationalUnpackedRemoteDataset
+
+__all__ = [
+    "HyperRelationalDataset",
+    "HyperRelationalUnpackedRemoteDataset",
+]
